Added new menu textures, changed some variables, updated settings menu. Current bug after clicking play game causes segfault Changes committed: 	new file:   Textures/Menu/False.png 	new file:   Textures/Menu/LeftArrow.png 	new file:   Textures/Menu/RelativeMovement.png 	new file:   Textures/Menu/RightArrow.png 	new file:   Textures/Menu/True.png 	modified:   include/Enemies/BasicEnemy.h 	modified:   include/Enemies/CorruptedStarShip.h 	modified:   include/Enemies/Enemy.h 	modified:   include/Enemies/HypnoEnemy.h 	modified:   include/GameTypes/MainMenu.h 	modified:   include/GameTypes/SettingsMenu.h 	new file:   include/GameTypes/Shop.h 	modified:   include/Menus/Button.h 	modified:   include/Menus/Label.h 	modified:   include/Mounts/BasicMount.h 	modified:   include/Mounts/HypnoMount.h 	modified:   include/Mounts/Mount.h 	modified:   include/Mounts/PurpleMount.h 	modified:   include/Mounts/TriangleMount.h 	modified:   include/Ships/FighterShip.h 	modified:   include/Ships/GalacticShip.h 	modified:   include/Ships/Ship.h 	modified:   include/Weapons/Weapon.h 	modified:   include/defines.h 	modified:   src/Enemies/BasicEnemy.cpp 	modified:   src/Enemies/CorruptedStarShip.cpp 	modified:   src/Enemies/HypnoEnemy.cpp 	modified:   src/GameTypes/Game.cpp 	modified:   src/GameTypes/MainMenu.cpp 	modified:   src/GameTypes/SettingsMenu.cpp 	new file:   src/GameTypes/Shop.cpp 	modified:   src/Menus/Button.cpp 	modified:   src/Menus/HighscoreScreen.cpp 	modified:   src/Menus/Label.cpp 	modified:   src/Mounts/BasicMount.cpp 	modified:   src/Mounts/HypnoMount.cpp 	modified:   src/Mounts/PurpleMount.cpp 	modified:   src/Mounts/TriangleMount.cpp 	modified:   src/Namespaces/Collisions.cpp 	modified:   src/Ships/FighterShip.cpp 	modified:   src/Ships/GalacticShip.cpp 	modified:   src/Weapons/PurplePlasma.cpp 	modified:   src/main.cpp

+++ include/Weapons/Weapon.h
@@ -17,8 +17,8 @@ class Weapon {
     
     bool getVisible() { return visible; }
     
-    int getDamage() { return damage; }
-    int hit() { visible = false; return damage; }
+    float getDamage() { return damage; }
+    float hit() { visible = false; return damage; }
     
     float getX() { return x; }
     float getY() { return y; }
@@ -26,7 +26,7 @@ class Weapon {
     float getHeight() { return height; }
     
   protected:
-    int damage;
+    float damage;
     float angle;
     bool visible;
     float directionX, directionY;
